requirements.txt
kafka-python
scikit-learn
pandas
numpy
pyspark

